% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_to_leaves.R
\name{apply_to_leaves}
\alias{apply_to_leaves}
\title{Add a list to the leaves of a nested list}
\usage{
apply_to_leaves(nestedList, toNest, pasteSep)
}
\arguments{
\item{nestedList}{The nested list whose leaves will be added to.}

\item{toNest}{The nested list to be added to each of the leaves.}

\item{pasteSep}{The character string that will be used to paste together list
names.}
}
\value{
A renamed nested list where the names describe the path to a list's
location within the nested list.
}
\description{
A function to added together nested lists, by adding one to
every leaf of another.
}
\examples{

vector1 <- c("a", "b", "c")
vector2 <- c("e", "f", "g", "h")
vector3 <- c("i", "j")

list1 <- vec_to_list(vector1)
list2 <- vec_to_list(vector2)
list3 <- vec_to_list(vector3)

nestedList <- build_nested_list(list1,
                                list2,
                                pasteSep = "__")

combinedList <- apply_to_leaves(nestedList = nestedList,
                                toNest = list3, pasteSep = "__")

}
\seealso{
prefix_all
}
