% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_available_area.R
\name{build_available_area}
\alias{build_available_area}
\title{Generate a polygon for use as an available area}
\usage{
build_available_area(
  movementData,
  method = c("MCP", "KDE_LSCV", "KDE_href", "AKDE", "dBBMM"),
  contour,
  SRS_string = "EPSG:32601"
)
}
\arguments{
\item{movementData}{must have a x and y column for locations, and a datetime column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}

\item{method}{"MCP", "KDE_LSCV", "KDE_href", "AKDE", "dBBMM"}

\item{contour}{Numeric, > 0 and < 100.}

\item{SRS_string}{= "EPSG:32601"}
}
\value{
A spatialpolygondataframe
}
\description{
A
}
