% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_brms.R
\name{diagnostics_brms}
\alias{diagnostics_brms}
\title{Generate plots to assess brms}
\usage{
diagnostics_brms(brmsResults)
}
\arguments{
\item{brmsResults}{output tar_targets resulting in areaBrms or ssfBrms}
}
\value{
a
}
\description{
A
}
