% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_estimates.R
\name{direct_estimates}
\alias{direct_estimates}
\title{Run all direct estimates at both scales}
\usage{
direct_estimates(simDataList)
}
\arguments{
\item{simDataList}{output from tar_combine}
}
\value{
a
}
\description{
A
}
