% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_estimate.R
\name{extract_estimate}
\alias{extract_estimate}
\title{Extract point estimates and SE}
\usage{
extract_estimate(OUT)
}
\arguments{
\item{OUT}{the output from any of the method functions}
}
\value{
A dataframe with the
}
\description{
A
}
