% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values_rmd.R
\name{extract_values_rmd}
\alias{extract_values_rmd}
\title{Generate files needed for RMD manuscript}
\usage{
extract_values_rmd()
}
\value{
Will save a series of csv files containing the raw estimates and
model betas that are needed to knit the manuscript.
}
\description{
Extract the required values to report in the RMD file.
}
