% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_effect_plots.R
\name{generate_effect_plots}
\alias{generate_effect_plots}
\title{Generate effect plots}
\usage{
generate_effect_plots(brmsResults, method)
}
\arguments{
\item{brmsResults}{output tar_targets resulting in ssfBrms or areaBrms}

\item{method}{c("rsf", "ssf", "wides")}
}
\value{
a
}
\description{
A
}
