% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_iteraction_plot.R
\name{generate_iteraction_plot}
\alias{generate_iteraction_plot}
\title{Generate interaction plot needed for RMD manuscript}
\usage{
generate_iteraction_plot()
}
\value{
Saves a png and pdf files.
}
\description{
Generate interaction plot needed for RMD manuscript
}
