% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spec_curves.R
\name{generate_spec_curves}
\alias{generate_spec_curves}
\title{Generate specification curves}
\usage{
generate_spec_curves(compiledResults, method)
}
\arguments{
\item{compiledResults}{output tar_targets resulting in areaCompiled or ssfCompiled}

\item{method}{c("rsf", "ssf", "wides")}
}
\value{
a
}
\description{
A
}
