% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_indi_rsf.R
\name{method_indi_rsf}
\alias{method_indi_rsf}
\title{Run RSF based analyses}
\usage{
method_indi_rsf(
  movementData,
  landscape,
  availableArea,
  availablePointsPer,
  weighting
)
}
\arguments{
\item{movementData}{must have a x and y column for locations, and a datetime column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}
}
\value{
a
}
\description{
A
}
