% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_indi_ssf.R
\name{method_indi_ssf}
\alias{method_indi_ssf}
\title{Run SSD based analyses}
\usage{
method_indi_ssf(
  movementData,
  landscape,
  methodForm,
  covExtract,
  availableSteps
)
}
\arguments{
\item{movementData}{must have a x and y column for locations, and a datetime column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}

\item{methodForm}{"mf.is", "mf.ss"}

\item{covExtract}{"start", "end", "both"}
}
\value{
a
}
\description{
A
}
