% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_indi_wides.R
\name{method_indi_wides}
\alias{method_indi_wides}
\title{Run wides based analyses}
\usage{
method_indi_wides(
  movementData,
  landscapeRaster,
  spSamp,
  availableArea,
  availablePointsPer,
  ...
)
}
\arguments{
\item{movementData}{must have a x and y column for locations, and a datetime column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}

\item{spSamp}{c("rd", "st")}
}
\value{
a
}
\description{
A
}
