% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_list.R
\name{nest_list}
\alias{nest_list}
\title{Nest one list within another}
\usage{
nest_list(parentList, toNest, pasteSep = "__", parentNames = names(parentList))
}
\arguments{
\item{parentList}{The list into which the second list will be nested within.}

\item{toNest}{The list that will be repeatedly added to every slot of the top
nest.}

\item{pasteSep}{The separator to use between level names during paste
process. Default is "__".}

\item{parentNames}{The names from the top list that will be pasted onto the
lower list names to mirror the path to the lower list slots.}
}
\value{
A named list, with the loweest level being replaced with the path.
}
\description{
Takes two named lists and nests the places the second within
every slot of the first list, while making the list names match the path
from the top list.
}
\details{
This function is only provided to aid with tree generation using the
build_nested_list function.
}
\examples{

vector1 <- c("a", "b", "c")
vector2 <- c("e", "f", "g", "h")

list1 <- vec_to_list(vector1)
list2 <- vec_to_list(vector2)

nest_list(parentList = list1, toNest = list2, pasteSep = "_")

}
\seealso{
build_nested_list
}
