% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_combined_results.R
\name{parse_combined_results}
\alias{parse_combined_results}
\title{Extract choices from combined results}
\usage{
parse_combined_results(combResults)
}
\arguments{
\item{combResults}{a}
}
\value{
a
}
\description{
A
}
