% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_landscape_example.R
\name{plot_landscape_example}
\alias{plot_landscape_example}
\title{Plot the landscapes used for simulation and the classified used for analysis}
\usage{
plot_landscape_example(species = "BADGER", seed = 2022)
}
\value{
a
}
\description{
A
}
