% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_all.R
\name{prefix_all}
\alias{prefix_all}
\title{Add prefix to a nested list}
\usage{
prefix_all(nestedList, prefix, pasteSep)
}
\arguments{
\item{nestedList}{A named nested list.}

\item{prefix}{A character string that will be applied to the names of a given
list.}

\item{pasteSep}{A character string that will be pasted between the names and
prefix.}
}
\value{
A renamed nested list.
}
\description{
A function to add a prefix to all named elements of a nested
list. Used with \code{apply_to_leaves()}.
}
\examples{

vector1 <- c("a", "b", "c")
vector2 <- c("e", "f", "g", "h")

list1 <- vec_to_list(vector1)
list2 <- vec_to_list(vector2)

nestedList <- build_nested_list(list1,
                                list2,
                                pasteSep = "__")

nestedListPrefix <- prefix_all(nestedList, "example", pasteSep = "__")

}
\seealso{
apply_to_leaves
}
