% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_brms.R
\name{run_brms}
\alias{run_brms}
\title{Run all brm models}
\usage{
run_brms(compiledResults, method)
}
\arguments{
\item{compiledResults}{output tar_combine resulting in ssfResults or areaResults}
}
\value{
a
}
\description{
A
}
