% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_individual.R
\name{simulate_individual}
\alias{simulate_individual}
\title{Simulate individual}
\usage{
simulate_individual(
  species,
  simSteps = 24 * 60 * 365,
  desOptions = 10,
  options = 12,
  landscapeList,
  seed
)
}
\arguments{
\item{species}{c("BADGER", "VULTURE", "KINGCOBRA")}

\item{simSteps}{= 24\*60 \*365}

\item{desOptions}{= 10}

\item{options}{= 12}
}
\value{
a
}
\description{
A
}
