% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_landscape.R
\name{simulate_landscape}
\alias{simulate_landscape}
\title{Simulate landscapes}
\usage{
simulate_landscape()
}
\arguments{
\item{movementData}{must have a x and y column for locations, and a datetime column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}
}
\value{
a
}
\description{
A
}
