% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty_vs_estimate.R
\name{uncertainty_vs_estimate}
\alias{uncertainty_vs_estimate}
\title{Generate uncertainty plots}
\usage{
uncertainty_vs_estimate(aResults, sResults)
}
\arguments{
\item{aResults}{output tar_targets resulting in areaResults}

\item{sResults}{output tar_targets resulting in ssfResults}
}
\value{
a
}
\description{
A
}
