% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_to_list.R
\name{vec_to_list}
\alias{vec_to_list}
\title{Convert a vector to a named list}
\usage{
vec_to_list(inVector)
}
\arguments{
\item{inVector}{A vector.}
}
\value{
A named list, with slots and names equal to the input vector.
}
\description{
A function to take a vector and convert it into a list where
each slot of the list contains a vector element and is named the same.
}
\examples{

vec_to_list(c("a", "b", "c"))

}
