% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_indi_area.R
\name{wrapper_indi_area}
\alias{wrapper_indi_area}
\title{Run all area based methods - wides and rsf}
\usage{
wrapper_indi_area(movementData, landscape, optionsList)
}
\arguments{
\item{movementData}{must have a x and y column for locations, and a datetime
column for timestamps ("\%Y-\%m-\%d \%H:\%M:\%S")}

\item{optionsList}{Must have the following items: Method_method areaMethod
areaContour Method_ap Method_sp Method_we}
}
\value{
a
}
\description{
A
}
